+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Version 0.8.0:
 - Calling "setOptions()" now only sets the ones you have passed (no longer reverts the others to their defaults).
+- Renamed the "Select Days To Exclude" dialog to "Repeat Options" (renames settings and CSS classes to match).
 
 ### Version 0.7.0:
 - Added Turn On/Off full-screen mode buttons to the toolbars (in any view).